Restructure parser project. Add comments and TODOs.
