--- src/main/java/de/tum/in/www1/bamboo/server/parser/ParserException.java
@@ -1,19 +1,0 @@
-package de.tum.in.www1.bamboo.server.parser;
-
-public class ParserException extends Exception {
-
-    public ParserException() {
-    }
-
-    public ParserException(String message) {
-        super(message);
-    }
-
-    public ParserException(Throwable cause) {
-        super(cause);
-    }
-
-    public ParserException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
