+++ src/main/java/de/tum/in/www1/bamboo/server/parser/exception/UnsupportedToolException.java
@@ -1,0 +1,8 @@
+package de.tum.in.www1.bamboo.server.parser.exception;
+
+public class UnsupportedToolException extends Exception {
+
+    public UnsupportedToolException(String message) {
+        super(message);
+    }
+}
